# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "The Tower API (v1)",
        "description": "Endpoints for various resources in The Tower.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
...
        "/menus/{id}": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "Get a menu",
                "description": "Retrieve a single menu.",
                "operationId": "3702f68603c17b4a04e5e5c35f31f5f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menus"
                ],
                "summary": "Delete menu",
                "description": "Soft-delete a menu.",
                "operationId": "60298d7d2d9100279e4904bc90ef15bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Menus"
                ],
                "summary": "Update menu",
                "description": "Update an existing menu.",
                "operationId": "e3ae92ff310432ca0784611b874a454b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMenuRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated menu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/menus": {
            "get": {
                "tags": [
                    "Menus"
                ],
                "summary": "List menus",
                "description": "List menus (paginated).",
                "operationId": "6a471d71b55cf04ae64844611e541c4a",
                "responses": {
                    "200": {
                        "description": "Paginated list of menus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuCollection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Menus"
                ],
                "summary": "Create menu",
                "description": "Create a new menu.",
                "operationId": "919180cbe58d90b236fa054287c4cfcc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMenuRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/menus/{id}/restore": {
            "patch": {
                "tags": [
                    "Menus"
                ],
                "summary": "Restore menu",
                "description": "Restore a soft-deleted menu.",
                "operationId": "1c62f9c9208d82ac548b16ea17872aae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restored menu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuResource"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Menu is not deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMenuRequest": {
                "description": "Payload sent when creating a new Menu",
                "required": [
                    "name",
                    "value"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "Epic Battle Theme"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.006
                    }
                },
                "type": "object"
            },
...
            "UpdateMenuRequest": {
                "description": "Payload for updating an existing Menu. All fields optional.",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "Epic Battle Theme Remix"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.007
                    }
                },
                "type": "object"
            },
...
            "MenuCollection": {
                "description": "Paginated list of Menus",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MenuResource"
                        }
                    },
                    "links": {
                        "properties": {
                            "first": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/menus?page=1"
                            },
                            "last": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/menus?page=2"
                            },
                            "prev": {
                                "type": "string",
                                "format": "uri",
                                "example": null,
                                "nullable": true
                            },
                            "next": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/menus?page=2",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "from": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 2
                            },
                            "path": {
                                "type": "string",
                                "format": "uri",
                                "example": "/api/v1/menus"
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 50
                            },
                            "to": {
                                "type": "integer",
                                "example": 50
                            },
                            "total": {
                                "type": "integer",
                                "example": 100
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "MenuResource": {
                "description": "Represents a Menu available in the API",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dark Being"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 0.006
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}